fix:stompdto so 내용 string으로 변환

+++ Backend/tranvel/src/main/java/com/ssafy/tranvel/entity/RandomGame.java
@@ -14,7 +14,7 @@ import lombok.NoArgsConstructor;
 public class RandomGame {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long miniGameCode;
+    private int miniGameCode;
 
     @Column(length = 16, name = "GameName")
     private String gameName;
